# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fitness AI API Documentation",
        "description": "L5 Swagger OpenApi description for the Fitness Backend",
        "contact": {
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrar novo usuário",
                "description": "Cria uma nova conta de usuário.",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "last_name",
                                    "email",
                                    "cpf",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João",
                                        "maxLength": 80
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Silva",
                                        "maxLength": 120
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@email.com",
                                        "maxLength": 180
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "123.456.789-00",
                                        "maxLength": 14
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "senha1234",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "senha1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário registrado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Autenticar usuário",
                "description": "Realiza login e retorna token de acesso.",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "senha1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso"
                    },
                    "422": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Revoga o token de acesso atual.",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Perfil do usuário autenticado",
                "description": "Retorna os dados do usuário autenticado, incluindo onboarding e gamificação.",
                "operationId": "9d9bc556f927575aa492a88c28fe5908",
                "responses": {
                    "200": {
                        "description": "Dados do usuário"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/onboarding": {
            "get": {
                "tags": [
                    "Onboarding"
                ],
                "summary": "Obter dados de onboarding",
                "description": "Retorna os dados de onboarding do usuário autenticado.",
                "operationId": "503081b518bc60df6e9b5762d473bc15",
                "responses": {
                    "200": {
                        "description": "Dados de onboarding"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Onboarding"
                ],
                "summary": "Enviar dados de onboarding",
                "description": "Salva as informações de onboarding do usuário.",
                "operationId": "64b8193d38da8e015281344ae9756359",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "gender": {
                                        "type": "string",
                                        "example": "M",
                                        "enum": [
                                            "M",
                                            "F",
                                            "other",
                                            "prefer_not_to_say"
                                        ]
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": 25,
                                        "maximum": 120,
                                        "minimum": 10
                                    },
                                    "height_cm": {
                                        "type": "integer",
                                        "example": 175,
                                        "maximum": 250,
                                        "minimum": 100
                                    },
                                    "weight_kg": {
                                        "type": "number",
                                        "example": 75.5,
                                        "maximum": 300,
                                        "minimum": 30
                                    },
                                    "body_fat_percent": {
                                        "type": "number",
                                        "example": 15,
                                        "maximum": 60,
                                        "minimum": 3
                                    },
                                    "workouts_per_week": {
                                        "type": "integer",
                                        "example": 4,
                                        "maximum": 7,
                                        "minimum": 0
                                    },
                                    "work_style": {
                                        "type": "string",
                                        "example": "white_collar",
                                        "enum": [
                                            "white_collar",
                                            "blue_collar",
                                            "sedentary",
                                            "moderate",
                                            "active"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Onboarding salvo com sucesso"
                    },
                    "400": {
                        "description": "Erro de validação"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}